backend/requirements.txt
fastApi
pydantic
uvicorn
cryptography
python-multipart
requests
pymysql
dotenv
python-jose
pyjwt
sqlalchemy
